fastapi-ai/requirements.txt
aioice==0.10.1
aiortc==1.13.0
annotated-types==0.7.0
anyio==4.10.0
av==14.4.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
contourpy==1.3.3
cryptography==45.0.7
cycler==0.12.1
dnspython==2.8.0
fastapi==0.116.1
filelock==3.19.1
fonttools==4.59.2
fsspec==2025.9.0
google-crc32c==1.7.1
h11==0.16.0
idna==3.10
ifaddr==0.2.0
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.2
matplotlib==3.10.6
mpmath==1.3.0
networkx==3.5
numpy==1.26.4
opencv-python-headless==4.9.0.80
packaging==25.0
pillow==11.3.0
polars==1.33.1
psutil==7.0.0
pycparser==2.23
pydantic==2.11.7
pydantic_core==2.33.2
pyee==13.0.0
pylibsrtp==0.12.0
pyOpenSSL==25.2.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
requests==2.32.5
scipy==1.16.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
starlette==0.47.3
sympy==1.14.0
typing-inspection==0.4.1
typing_extensions==4.15.0
ultralytics==8.3.197
ultralytics-thop==2.0.17
urllib3==2.5.0
uvicorn==0.35.0
websockets==15.0.1